+++ test/spec/Feature/OpenApi/DisabledOpenApiSpec.hs
@@ -17,4 +17,5 @@ spec =
         [("Accept","application/openapi+json")] ""
         `shouldRespondWith`
         [json| {"code":"PGRST126","details":null,"hint":null,"message":"Root endpoint metadata is disabled"} |]
-        { matchStatus = 404 }
+        { matchStatus = 404
+        , matchHeaders = ["Content-Length" <:> "93"]}